add max min avg evaluator

+++ bee-engine/src/main/java/com/dianping/bee/engine/evaluator/function/CountEvaluator.java
@@ -4,15 +4,15 @@ import com.alibaba.cobar.parser.ast.expression.primary.function.groupby.Count;
 import com.dianping.bee.engine.evaluator.AbstractEvaluator;
 import com.dianping.bee.engine.spi.RowContext;
 
-public class CountEvaluator extends AbstractEvaluator<Count, Integer> {
+public class CountEvaluator extends AbstractEvaluator<Count, Long> {
 	public static final String ID = Count.class.getName();
 
-	private int m_count;
+	private long m_count;
 
 	@Override
-	public Integer evaluate(RowContext ctx, Count expr) {
+	public Long evaluate(RowContext ctx, Count expr) {
 		m_count++;
-		return 0;
+		return 0L;
 	}
 
 	@Override
@@ -22,7 +22,7 @@ public class CountEvaluator extends AbstractEvaluator<Count, Integer> {
 
 	@Override
 	public Class<?> getResultType(Count expr) {
-		return Integer.class;
+		return Long.class;
 	}
 
 	@Override
